MOOC/C9.x/C9-2-2: Use size_t in function mylen()

+++ MOOC/C9.x/C9-2-2-mylen.c
@@ -9,9 +9,9 @@
 #include "C9-2-2.h"
 
 
-int mylen(const char *s)
+size_t mylen(const char *s)
 {
-    int cnt = 0;
+    size_t cnt = 0;
     while (*(s+cnt) != '\0') {
         cnt ++;
     }
