+++ MOOC/C9.x/C9-2-2.h
@@ -15,4 +15,4 @@
 #include <stdio.h>
 #include <string.h>
 
-int mylen(const char *s);
+size_t mylen(const char *s);
